refactor operators into files
